populate HouseAndLand

+++ Land.java
@@ -3,10 +3,10 @@ public class Land {
     // Land Instance variables
     private int lotNumber;
     private String address;
-    private float landArea; // landArea to be in meters squared
+    private double landArea; // landArea to be in meters squared
 
     // Land constructor
-    public Land(int lotNumber, String address, float landArea) {
+    public Land(int lotNumber, String address, double landArea) {
         this.lotNumber = lotNumber;
         this.address = address;
         this.landArea = landArea;
@@ -38,7 +38,7 @@ public class Land {
     }
 
     // landArea getter
-    public float getLandArea() {
+    public double getLandArea() {
         return landArea;
     }
